% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgsvd.R
\name{cgsvd}
\alias{cgsvd}
\title{CGSVD}
\usage{
cgsvd(
  Y,
  X,
  c1,
  c2,
  R = 2,
  init = "svd",
  eps.pi = 1e-12,
  itermax.pi = 500,
  eps.pocs = 1e-12,
  itermax.pocs = 500,
  v.partition = F,
  Gcol = NULL,
  Grow = NULL
)
}
\arguments{
\item{Y}{A data frame or matrix with factors}

\item{c1}{The radius of l1 ball (constraint)}

\item{c2}{The radius of l2 ball (constraint)}

\item{R}{The number of singular triplets we want to obtain}

\item{init}{Method to initialize the singular vectors, by default "gsvd" (can also be "rand")}

\item{eps.pi}{Precision}

\item{itermax.pi}{The maximum number of iteration of power iteration}

\item{eps.pocs}{Precision}

\item{itermax.pocs}{The maximum number of iteration of projections}

\item{v.partition}{If true, creates a group constraint based on the ariables partition}

\item{Gcol}{A group partition of the data columns (categories)}

\item{Grow}{A group partition of the data rows}
}
\value{
Returns the constrained singular triplets of factor data and the number of iteration used to obtain them.
}
\description{
CGSVD
}
\examples{
library(FactoMineR)
data(tea)
tea <- tea [, 1:8]

test <- cgsvd (tea, 1.5, 1.5, R = 5, init = "rand")
}
