% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvd2.R
\name{csvd}
\alias{csvd}
\title{CSVD}
\usage{
csvd(
  X,
  c1,
  c2,
  init = "svd",
  R = 3,
  eps.pi = 1e-12,
  itermax.pi = 1000,
  eps.pocs = 1e-12,
  itermax.pocs = 1000,
  Gcol = NULL,
  Grow = NULL
)
}
\arguments{
\item{X}{A matrix of numerics}

\item{c1}{The radius of l1 ball (constraint)}

\item{c2}{The radius of l2 ball (constraint)}

\item{init}{Method to initialize the singular vectors, by default "svd" (can also be "rand")}

\item{R}{The number of singular triplets we want to obtain}

\item{eps.pi}{Precision}

\item{itermax.pi}{The maximum number of iteration of power iteration}

\item{eps.pocs}{Precision}

\item{itermax.pocs}{The maximum number of iteration of projections}

\item{Gcol}{A group partition of the data rows}

\item{Grow}{A group partition of the data columns}
}
\value{
Returns the constrained singular triplets of numeric data and the number of iteration used to obtain them
}
\description{
CSVD
}
\examples{
library(FactoMineR)
data("decathlon")
deca <- as.matrix(decathlon[,-13])
c1 <- sqrt(nrow(deca))/2
c2 <- sqrt(ncol(deca))/2
csvd2(deca, c1, c2, R=2)
}
