% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V_subvector.R
\name{foo_tmp}
\alias{foo_tmp}
\title{Create a vector of B12 inter B2 projection}
\usage{
foo_tmp(g, x)
}
\arguments{
\item{X}{A numeric data matrix issued from a qualitative data}

\item{G}{A partition of the matrix columns}
}
\value{
Returns the vector v containing the norm2 of each group.
}
\description{
Create a vector of B12 inter B2 projection
}
\examples{
G <- list(1, c(2, 3), c(4, 5))
a <- rnorm(5)
v_subvector(a, G)

}
