% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpmd.R
\name{gpmd}
\alias{gpmd}
\title{Generalized PMD}
\usage{
gpmd(Y, X, sumabsu, sumabsv, K = 3)
}
\arguments{
\item{Y}{A data frame or matrix with factors}

\item{sumabsu}{The constraint on the rows (numeric)}

\item{sumabsv}{The constraint on the columns (numeric)}

\item{K}{The number of singular triplets we want to obtain}
}
\value{
Returns the constrained singular triplets of factor data with the penalised matrix decomposition from the package PMA.
}
\description{
Generalized PMD
}
\examples{
library(PMD)
library(FactoMineR)
data(tea)
tea <- tea [, 1:8]

test <- gpmd (tea, 1.5, 1.5, K = 5)
}
