% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsvd.R
\name{gsvd}
\alias{gsvd}
\title{GSVD}
\usage{
gsvd(Y, X, R = 5)
}
\arguments{
\item{Y}{A data frame or matrix with factors}

\item{R}{The number of components}
}
\value{
Returns the R first generalized singular triplets (U, V, d) of a data frame with factors
}
\description{
GSVD
}
\examples{
library(FactoMineR)
data(tea)
tea <- tea [, 1:8]
gsvd(tea, 3)
}
