% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCA.R
\name{MCA}
\alias{MCA}
\title{Multivariate Correspondance Analysis}
\usage{
MCA(Y, n = 5)
}
\arguments{
\item{Y}{A data frame or matrix with factors}

\item{n}{The number of components}
}
\value{
Returns the R first factor scores of Y
}
\description{
Multivariate Correspondance Analysis
}
\examples{
data(cheese)
MCA(cheese, 3)
}
