% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm1.R
\name{norm1}
\alias{norm1}
\title{Norm1}
\usage{
norm1(vect)
}
\arguments{
\item{vect}{a vector}
}
\value{
the l1-norm of a vector
}
\description{
This function returns the l1 norm of a vector.
}
\examples{
norm1 (rnorm(10))
}
