% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Return the normalization of a vector (the vector divided by its l2 norm)
}
\description{
Normalize a vector
}
\examples{
normalize (rnorm(10))

}
