% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Principal Composant analysis}
\usage{
PCA(X, n = 5)
}
\arguments{
\item{X}{A data frame or matrix with numerics}

\item{n}{The number of components we want to obtain}
}
\value{
Returns the eigen values, factor scores ad csvd of a principal components analysis.
}
\description{
Principal Composant analysis
}
\examples{
library(FactoMineR)
data(decathlon)
deca <- scale(decathlon[, -13], center=T)/sqrt(nrow(decathlon)-1)
PCA(deca, n = 4)
}
