% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI1.R
\name{PI.1}
\alias{PI.1}
\title{Power iteration first component}
\usage{
PI.1(X, c1, c2, U0, V0, itermax = 1000, eps = 1e-16)
}
\arguments{
\item{X}{A matrix of numerics}

\item{c1}{the radius of l1 ball}

\item{c2}{the radius of l2 ball}

\item{U0}{Initial U values (random numeric vector or SVD first component)}

\item{V0}{Initial V values (random numeric vector or SVD first component)}

\item{itermax}{The maximum number of iteration}

\item{eps}{Precision}
}
\value{

}
\description{
Power iteration first component
}
\examples{

}
