% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.R
\name{PI}
\alias{PI}
\title{Power iteration}
\usage{
PI(
  X,
  c1,
  c2,
  P0,
  Q0,
  P = NULL,
  Q = NULL,
  eps.pi = 1e-16,
  itermax.pi = 1000,
  eps.pocs = 1e-16,
  itermax.pocs = 1000,
  Gcol = NULL,
  Grow = NULL
)
}
\arguments{
\item{X}{A matrix of numerics}

\item{c1}{the radius of l1 ball}

\item{c2}{the radius of l2 ball}

\item{P0}{Initial U values (random numeric vector or SVD first component)}

\item{Q0}{Initial V values (random numeric vector or SVD first component)}

\item{P}{A matrix containing the previous singular values U}

\item{Q}{A matrix containing the previous singular values V}

\item{itermax}{The maximum number of iteration}

\item{eps}{Precision}
}
\value{
Returns the singular triplet of a matrix, given the already found singular triplets.
}
\description{
Power iteration
}
\examples{
data(cheese)
X <- tab_disjonctif(cheese)
c1 <- 4
c2 <- 1.5
P0 <- normalize(rnorm(dim(X)[1]))
Q0 <- normalize(rnorm(dim(X)[2]))
PI (X, c1, c2, P0, Q0)
}
