% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.l1.R
\name{proj.l1}
\alias{proj.l1}
\title{Projection into l1 Ball}
\usage{
\method{proj}{l1}(x, c)
}
\arguments{
\item{x}{A vector.}

\item{c}{A constraint (numeric).}
}
\value{
Returns the projection of vector x into B1(c).
}
\description{
Projection into l1 Ball
}
\examples{
proj.l1(rnorm(10), 1.5)
}
