% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.l1l2.lambda.R
\name{proj.l1l2.lambda}
\alias{proj.l1l2.lambda}
\title{Lambda argument of the projection into the intersection of l1 and l2 Balls}
\usage{
\method{proj}{l1l2.lambda}(x, c)
}
\arguments{
\item{x}{A vector}

\item{c}{A constraint (numeric)}
}
\value{
the lamdba argument of the proximal operator
}
\description{
Lambda argument of the projection into the intersection of l1 and l2 Balls
}
\examples{
proj.l1l2.lambda(rnorm(10), 1.5)
}
