% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.l1l2.orth.R
\name{proj.l1l2.orth}
\alias{proj.l1l2.orth}
\title{Projection into intersection orthonormal matrix and l1 and l2 balls}
\usage{
\method{proj}{l1l2.orth}(x, c, M = NULL, itermax.pocs = 1000, eps.pocs = 1e-16, G = NULL)
}
\arguments{
\item{x}{A vector.}

\item{c}{A constraint (numeric).}

\item{M}{A matrix.}

\item{G}{A group partition of the vector x}
}
\value{
The projected vector x into the intersection of the orthonormal space of M and the l1 and l2 Balls intersection.
}
\description{
Projection into intersection orthonormal matrix and l1 and l2 balls
}
\examples{

}
