% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.l1l2.R
\name{proj.l1l2}
\alias{proj.l1l2}
\title{Projection into the intersection of l1 and l2 Balls}
\usage{
\method{proj}{l1l2}(x, c)
}
\arguments{
\item{x}{A vector.}

\item{c}{A constraint (numeric).}
}
\value{
Returns the vector x projected into B1(c)interB2(1).
}
\description{
Projection into the intersection of l1 and l2 Balls
}
\examples{
proj.l1l2(rnorm(10), 1.5)


}
