% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.lgl2.R
\name{proj.lgl2}
\alias{proj.lgl2}
\title{The projection into the intersection of l12 and l2 Balls}
\usage{
\method{proj}{lgl2}(x, c, G)
}
\arguments{
\item{x}{A vector}

\item{c}{A constraint (numeric)}

\item{G}{A group partition of the vector x}
}
\value{
Returns the vector x projected into B12(c)interB2(1).
}
\description{
The projection into the intersection of l12 and l2 Balls
}
\examples{
G <- list(1:3, 4:5)
x <- rnorm(5)
c <- 2*sqrt(5)/3
proj.l12l2(x, c, G)
}
