% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.orth.R
\name{proj.orth}
\alias{proj.orth}
\title{Orthonormal projection}
\usage{
\method{proj}{orth}(x, MtM)
}
\arguments{
\item{x}{A vector to project.}

\item{M}{A matrix. The vector is projected into the orthonormal space of M. #the matrix into which orthonormal space the vector is projected.}
}
\value{
The vector x projected into the orthonormal space of M.
}
\description{
Orthonormal projection
}
\examples{
x <- rnorm(10)
A <- matrix(data = c(1, 0.2, -0.3, 0.3, 0.5), nrow=10, ncol=3)
proj.orth(x, A)
}
