% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi.R
\name{psi}
\alias{psi}
\title{Intermediate Function of proj.l1l2}
\usage{
psi(x, lambda)
}
\arguments{
\item{x}{A vector}

\item{lambda}{Argument}
}
\value{

}
\description{
Intermediate Function of proj.l1l2
}
\examples{

}
