% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot_new.R
\name{screeplot_new}
\alias{screeplot_new}
\title{Gives the Screeplot of a Sparce MCA or PCA}
\usage{
screeplot_new(
  res,
  aff.mean = F,
  sort.comp = T,
  title = "Scree plot",
  ncomp = "all"
)
}
\arguments{
\item{res}{the result of SMCA or SPCA}

\item{aff.mean}{If TRUE, adds an horizontal line corresponding to the mean of the eigen values.}

\item{sort.comp}{If TRUE (default), the components are sorted by decreasing size.}
}
\value{
A ggplot2 object
}
\description{
Gives the Screeplot of a Sparce MCA or PCA
}
\examples{
data(cheese)
res.gpmd <- SMCA(cheese, n = 2, c1 = 3, c2 = 1.9, meth = "gpmd")
screeplot_new(res.gpmd)
}
