% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMCA.R
\name{SMCA}
\alias{SMCA}
\title{Sparse Multivariate Correspondance Analysis}
\usage{
SMCA(
  Y,
  c1,
  c2,
  n = 5,
  meth = "cgsvd",
  init = "rand",
  v.partition = F,
  Grow = NULL,
  Gcol = NULL
)
}
\arguments{
\item{Y}{A data frame or matrix with factors}

\item{c1}{The radius of l1 ball (constraint). It can take values from 1 to the square root of the number of row. The closer to 1 it is, the more sparsified are the observations.}

\item{c2}{The radius of l2 ball (constraint). It can take values from 1 to the square root of the number of columns. The closer to 1 it is, the more sparsified are the variables.}

\item{n}{The number of components we want to obtain}

\item{meth}{Method to create the sparse singular vectors, by default "cgsvd" (can also be "gpmd")}

\item{init}{Method to initialize the singular vectors, by default "rand" (can also be "svd")}

\item{v.partition}{If true, creates a group constraint based on the ariables partition (by default False)}

\item{Grow}{A group partition of the data row}

\item{Gcol}{A group partition of the data columns (categories)}
}
\value{
Returns the sparse MCA for the given constraints.
}
\description{
Sparse Multivariate Correspondance Analysis
}
\examples{
data(cheese)
SMCA(cheese, c1 = sqrt(nrow(cheese))/2, c2 = sqrt(ncol(cheese))/2, n = 4)
}
