% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft threshold.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Soft Threshold norme 1}
\usage{
soft_threshold(x, alpha)
}
\arguments{
\item{x}{a vector}

\item{alpha}{a numeric threshold}
}
\value{
Returns the proximal vector of x for a threshold alpha.
}
\description{
Soft Threshold norme 1
}
\examples{
#the proximal vector of a random vector
soft_threshold(rnorm(10), 0.2)
}
