% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft threshold normG.R
\name{soft_threshold_normg}
\alias{soft_threshold_normg}
\title{Soft Threshold norme de groupe}
\usage{
soft_threshold_normg(x, alpha, G)
}
\arguments{
\item{x}{a vector}

\item{alpha}{a numeric threshold}

\item{G}{A group partition of the vector x}
}
\value{
Returns the group norm proximal vector of x for a threshold alpha.
}
\description{
Soft Threshold norme de groupe
}
\examples{
#the proximal vector of a random vector
G <- list(1:4, 5, 6:8, 9:10)
x <- rnorm(10)
c <- 1.5
lambda <- proj.l1l2.lambda(x, c)
soft_threshold_normg(x, lambda, G)
}
