% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPCA.R
\name{SPCA}
\alias{SPCA}
\title{Sparse Principal Components Analysis}
\usage{
SPCA(X, c1, c2, n)
}
\arguments{
\item{X}{A data frame or matrix with numerics}

\item{c1}{The radius of l1 ball (constraint). It can take values from 1 to the square root of the number of row. The closer to 1 it is, the more sparsified are the observations.}

\item{c2}{The radius of l2 ball (constraint). It can take values from 1 to the square root of the number of columns. The closer to 1 it is, the more sparsified are the variables.}

\item{n}{The number of components we want to obtain}
}
\value{
Returns the eigen values, factor scores ad csvd of a sparse principal components analysis.
}
\description{
Sparse Principal Components Analysis
}
\examples{
library(FactoMineR)
data(decathlon)
deca <- scale(decathlon[, -13], center=T)/sqrt(nrow(decathlon)-1)
SPCA(deca, c1 = sqrt(nrow(deca))/2, c2 = sqrt(ncol(deca))/2, n = 4)
}
